refactor(eve): adapt task schema cleanup to canonical sources
